review changes

+++ backend/api/views.py
@@ -2,6 +2,7 @@ from django.db.models import Exists, OuterRef, Sum
 from django.db.models.functions import Lower
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404, redirect
+from rest_framework import generics
 from rest_framework import status, viewsets
 from rest_framework.decorators import action
 from rest_framework.exceptions import ValidationError
@@ -9,7 +10,7 @@ from rest_framework.generics import (
     ListAPIView,
 )
 from rest_framework.permissions import (
-    IsAuthenticated, IsAuthenticatedOrReadOnly,
+    IsAuthenticated, IsAuthenticatedOrReadOnly, AllowAny
 )
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -19,12 +20,12 @@ from .models import (
     ShoppingCart, Subscription, Tag, User,
 )
 from .pagination import CustomPagination
-from .permissions import IsAdminOrReadOnly, IsAuthor
+from .permissions import IsAuthor
 from .serializers import (
     AvatarSerializer, IngredientSerializer,
     RecipeCreateSerializer, RecipeListSerializer,
     RecipeMinifiedSerializer, TagSerializer,
-    UserWithRecipesSerializer,
+    UserWithRecipesSerializer, CustomUserSerializer
 )
 
 
@@ -33,17 +34,17 @@ def redirect_to_recipe(request, code):
     return redirect(f'/recipes/{recipe.id}/')
 
 
-class TagViewSet(viewsets.ModelViewSet):
+class TagViewSet(viewsets.ReadOnlyModelViewSet):
     queryset = Tag.objects.all()
     serializer_class = TagSerializer
-    permission_classes = (IsAdminOrReadOnly,)
+    permission_classes = (AllowAny,)
     pagination_class = None
     lookup_field = 'id'
 
 
-class IngredientViewSet(viewsets.ModelViewSet):
+class IngredientViewSet(viewsets.ReadOnlyModelViewSet):
     serializer_class = IngredientSerializer
-    permission_classes = (IsAdminOrReadOnly,)
+    permission_classes = (AllowAny,)
     pagination_class = None
     lookup_field = 'id'
 
@@ -74,7 +75,33 @@ class RecipeViewSet(viewsets.ModelViewSet):
         return super().get_permissions()
 
     def perform_create(self, serializer):
+        tags = self.request.data.get('tags', [])
+        if len(tags) != len(set(tags)):
+            from rest_framework.exceptions import ValidationError
+            raise ValidationError({'tags': 'Теги не должны повторяться'})
         serializer.save(author=self.request.user)
+
+    def partial_update(self, request, *args, **kwargs):
+        recipe = get_object_or_404(Recipe, pk=kwargs.get('pk'))
+        if recipe.author != request.user:
+            return Response(
+                {'errors': 'Вы не являетесь автором этого рецепта'},
+                status=status.HTTP_403_FORBIDDEN
+            )
+        tags = request.data.get('tags', [])
+        if len(tags) != len(set(tags)):
+            return Response(
+                {'errors': 'Теги не должны повторяться'},
+                status=status.HTTP_400_BAD_REQUEST
+            )
+        serializer = self.get_serializer(
+            recipe,
+            data=request.data,
+            partial=True
+        )
+        serializer.is_valid(raise_exception=True)
+        self.perform_update(serializer)
+        return Response(serializer.data)
 
     def get_queryset(self):
         queryset = super().get_queryset()
@@ -255,16 +282,25 @@ class SubscribeView(APIView):
                 status=status.HTTP_400_BAD_REQUEST,
             )
         Subscription.objects.create(user=user, author=author)
-        serializer = UserWithRecipesSerializer(
-            author, context={'request': request}
-        )
+        recipes_limit = request.query_params.get('recipes_limit')
+        context = {'request': request}
+        if recipes_limit:
+            try:
+                context['recipes_limit'] = int(recipes_limit)
+            except (ValueError, TypeError):
+                pass
+        serializer = UserWithRecipesSerializer(author, context=context)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
     def delete(self, request, id):
         author = get_object_or_404(User, id=id)
-        user = request.user
+        if author == request.user:
+            return Response(
+                {'errors': 'Нельзя отписаться от самого себя'},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
         deleted, _ = Subscription.objects.filter(
-            user=user, author=author
+            user=request.user, author=author
         ).delete()
         if not deleted:
             return Response(
@@ -272,6 +308,29 @@ class SubscribeView(APIView):
                 status=status.HTTP_400_BAD_REQUEST,
             )
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+    def partial_update(self, request, *args, **kwargs):
+        """Обновление рецепта (частичное)"""
+        recipe = self.get_object()
+        if recipe.author != request.user:
+            return Response(
+                {'errors': 'Вы не являетесь автором этого рецепта'},
+                status=status.HTTP_403_FORBIDDEN
+            )
+        tags = request.data.get('tags', [])
+        if len(tags) != len(set(tags)):
+            return Response(
+                {'errors': 'Теги не должны повторяться'},
+                status=status.HTTP_400_BAD_REQUEST
+            )
+        serializer = self.get_serializer(
+            recipe,
+            data=request.data,
+            partial=True
+        )
+        serializer.is_valid(raise_exception=True)
+        self.perform_update(serializer)
+        return Response(serializer.data)
 
 
 class AvatarView(APIView):
@@ -297,3 +356,21 @@ class AvatarView(APIView):
             user.avatar = None
             user.save()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+class UserDetailView(generics.RetrieveAPIView):
+    """
+    Получение пользователя по ID с его рецептами.
+    """
+    queryset = User.objects.all()
+    serializer_class = CustomUserSerializer
+    permission_classes = []
+
+
+class UserMeView(generics.RetrieveAPIView):
+    queryset = User.objects.all()
+    serializer_class = CustomUserSerializer
+    permission_classes = (IsAuthenticated,)
+
+    def get_object(self):
+        return self.request.user
